+++ eslint-plugin-agama-i18n.js
@@ -6,6 +6,7 @@
 const stringLiteralsRule = require("./string-literals");
 const topLevelRule = require("./top-level-translation");
 const multipleSpaceRule = require("./multiple-space");
+const markedString = require("./marked-string");
 
 module.exports = {
   rules: {
@@ -13,5 +14,6 @@ module.exports = {
     "string-literals": stringLiteralsRule,
     "top-level-translation": topLevelRule,
     "multiple-space": multipleSpaceRule,
+    "marked-string": markedString,
   },
 };